checkout if anything needs to be deleted

+++ vars/dockerCompose.groovy
@@ -2,7 +2,9 @@ def up(String file) {
 	stage("Docker Up") {
 		bash "docker-compose -f ${file} down"
 		volumes = sh(script: "docker volume ls -qf dangling=true", returnStdout: true)
-		bash "docker volume rm ${volumes}"
+		if (!volumes.trim()) {
+			bash "docker volume rm ${volumes}"
+		}
 		bash "docker-compose -f ${file} up -d --build"
 	}
 }
